Adding api access (#3)
* Refactors for configuring the services, the pipeline middlewares and the endpoints.

* Adding Swagger documentation for the weather forecast endpoint using the Produces methods.
